anytime_gbfs: completed
Testing with Manhattan distance heuristic:

Of 20 initial problems, 13 were solved in less than 5 seconds by this solver.
Of the 13 problems that were solved, the cost of 11 matched or outperformed the benchmark.
Problems that remain unsolved in the set are Problems: [3, 5, 13, 14, 16, 17, 19]
The benchmark implementation solved 15 out of the 20 practice problems given 5 seconds.